% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{geocodeAddress}
\alias{geocodeAddress}
\title{geocodeAddress}
\usage{
geocodeAddress(address, api_key)
}
\arguments{
\item{address}{Endereco de forma descrita}

\item{api_key}{Chave de acesso da API do googlemaps}
}
\description{
Retorna a latitude e longitude de um determinado endereco com base na API do googlemaps.
}
