% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{geocodeDist}
\alias{geocodeDist}
\title{geocodeDist}
\usage{
geocodeDist(origin, destination, mode, api_key)
}
\arguments{
\item{origin}{Latutide e longitude do endereço de origem}

\item{destination}{Latutide e longitude do endereço de destino}

\item{mode}{Forma de deslocamento; \code{"driving"}, \code{'walking'}, \code{'bicycling'} or \code{'transit'}}

\item{api_key}{Chave de acesso da API do googlemaps}
}
\description{
Retorna a distância e o tempo de deslocamento entre dois endereços
}
