% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{geocodeLatlong}
\alias{geocodeLatlong}
\title{geocodeLatlong}
\usage{
geocodeLatlong(latlong, api_key)
}
\arguments{
\item{latlong}{Latutide e longitude do endereço}

\item{api_key}{Chave de acesso da API do googlemaps}
}
\description{
Retorna o endereço através da latitude e longitude
}
